add second task

+++ DSP-lab1/src/by/bsuir/dsp/lab/Functions.java
@@ -1,8 +1,8 @@
 package by.bsuir.dsp.lab;
 
 public class Functions {
-	
-	public static double func1(int A, int f, int n, double fis) {
-		 return A * Math.sin(2 * Math.PI * f * n / Constants.N + fis);
+
+	public static double func1(double A, double f, int n, double fis) {
+		return A * Math.sin(2 * Math.PI * f * n / Constants.N + fis);
 	}
 }
